+++ frontend/src/components/AlertsList.vue
@@ -1,119 +1,44 @@
 <script setup>
-import axios from "axios";
-import { ref, onMounted, computed } from "vue";
+import { computed, defineProps } from "vue";
 import { useStore } from "vuex";
-import { useToast } from "vue-toastification";
+import { defineEmits } from "vue";
+
+const props = defineProps({
+  alerts: {
+    type: Array,
+    default: () => [],
+  },
+})
 
 const store = useStore();
-const alerts = ref([]);
-const offset = ref(0);
-const limit = 10;
-const isLoading = ref(false);
 const isAuthenticated = computed(() => store.getters.isAuthenticated);
-const toast = useToast();
+const emit = defineEmits(['refresh', 'loadMore', 'ack']);
 
-function refresh() {
-  isLoading.value = true;
-  offset.value = 0;
-  
-  axios
-    .get("/refresh")
-    .then(() => {
-      loadAlerts();
-      toast.success("Alerts loaded");
-    })
-    .catch((error) => {
-      if (error.response.status === 401) {
-        isLoading.value = false;
-        store.dispatch("logout");
-        toast.error("Session expired");
-      } else {
-        toast.error("Error loading alerts");
-      }
-    });
+function emitRefreshEvent() {
+  emit('refresh');
 }
 
-
-function loadAlerts() {
-  axios
-    .get(`/alerts/${offset.value}`)
-    .then((response) => {
-      const newAlerts = response.data.map(formatAlert);
-      if (offset.value === 0) {
-        alerts.value = newAlerts;
-      } else {
-        alerts.value = [...alerts.value, ...newAlerts];
-      }
-      offset.value += limit;
-    })
-    .catch((error) => {
-      toast.error("Error loading alerts");
-    })
-    .finally(() => {
-      isLoading.value = false;
-    });
+function emitLoadMoreEvent() {
+  emit('loadMore');
 }
 
-function formatAlert(alert) {
-  const inputDate = new Date(alert.created_at);
-  const day = inputDate.getDate().toString().padStart(2, "0");
-  const month = (inputDate.getMonth() + 1).toString().padStart(2, "0");
-  const year = inputDate.getFullYear().toString().slice(-2);
-  const hours = inputDate.getHours().toString().padStart(2, "0");
-  const minutes = inputDate.getMinutes().toString().padStart(2, "0");
-  const formattedDate = `${day}.${month}.${year} ${hours}:${minutes}`;
-  return { ...alert, created_at: formattedDate };
+function emitAckEvent(alertId) {
+  emit('ack', alertId);
 }
-
-function loadMore() {
-  loadAlerts();
-}
-
-function ackAlert(alertId) {
-  axios
-    .get(`/alert/${alertId}/ack`, {
-      headers: { Authorization: `Bearer ${store.getters.getToken}` },
-    })
-    .then(() => {
-      const alertIndex = alerts.value.findIndex(
-        (alert) => alert.id === alertId
-      );
-      if (alertIndex !== -1) {
-        alerts.value[alertIndex].is_acked = !alerts.value[alertIndex].is_acked;
-      }
-    })
-    .catch((error) => {});
-}
-
-onMounted(() => {
-  loadAlerts();
-});
 </script>
 
 <template>
   <section>
-    <div class="isLoading" v-if="isLoading">
-      <div class="lds-roller">
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-        <div></div>
-      </div>
-    </div>
     <header>
       <h2>Alerts</h2>
-      <button v-if="isAuthenticated" class="refresh" @click="refresh">Refresh</button>
+      <button v-if="isAuthenticated" class="refresh" @click="emitRefreshEvent">Refresh</button>
     </header>
     <ul
       class="alerts"
       v-infinite-scroll="loadMore"
       infinite-scroll-distance="200"
     >
-      <li v-if="!isLoading" v-for="alert in alerts" :key="alert.id" 
+      <li v-for="alert in alerts" :key="alert.id" 
           :class="{'alert': true, 'acked': isAuthenticated && alert.is_acked}">
 
         <header>
@@ -125,7 +50,7 @@ onMounted(() => {
             <p class="time">{{ alert.created_at }}</p>
           </div>
 
-          <button v-if="isAuthenticated" @click="() => ackAlert(alert.id)">
+          <button v-if="isAuthenticated" @click="() => emitAckEvent(alert.id)">
             <span v-if="!alert.is_acked">ACK</span>
             <span v-if="alert.is_acked">UNACK</span>
           </button>
@@ -148,7 +73,7 @@ onMounted(() => {
         </footer>
       </li>
     </ul>
-    <button v-if="isAuthenticated" class="loadMore" @click="loadMore">Load more</button>
+    <button v-if="isAuthenticated" class="loadMore" @click="emitLoadMoreEvent">Load more</button>
   </section>
 </template>
 
@@ -170,107 +95,6 @@ section {
   @media (min-width: 768px) {
     height: 100%;
     min-height: 70vh;
-  }
-
-  .isLoading {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    min-height: 100vh;
-    border-radius: 5px;
-    background: $background;
-    border: 1px solid $lighter-background;
-    display: flex;
-    justify-content: center;
-    padding-top: 10rem;
-    box-sizing: border-box;
-    z-index: 9;
-
-    .lds-roller {
-      display: inline-block;
-      position: relative;
-      width: 80px;
-      height: 80px;
-    }
-    .lds-roller div {
-      animation: lds-roller 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-      transform-origin: 40px 40px;
-    }
-    .lds-roller div:after {
-      content: " ";
-      display: block;
-      position: absolute;
-      width: 7px;
-      height: 7px;
-      border-radius: 50%;
-      background: $darker-foreground;
-      margin: -4px 0 0 -4px;
-    }
-    .lds-roller div:nth-child(1) {
-      animation-delay: -0.036s;
-    }
-    .lds-roller div:nth-child(1):after {
-      top: 63px;
-      left: 63px;
-    }
-    .lds-roller div:nth-child(2) {
-      animation-delay: -0.072s;
-    }
-    .lds-roller div:nth-child(2):after {
-      top: 68px;
-      left: 56px;
-    }
-    .lds-roller div:nth-child(3) {
-      animation-delay: -0.108s;
-    }
-    .lds-roller div:nth-child(3):after {
-      top: 71px;
-      left: 48px;
-    }
-    .lds-roller div:nth-child(4) {
-      animation-delay: -0.144s;
-    }
-    .lds-roller div:nth-child(4):after {
-      top: 72px;
-      left: 40px;
-    }
-    .lds-roller div:nth-child(5) {
-      animation-delay: -0.18s;
-    }
-    .lds-roller div:nth-child(5):after {
-      top: 71px;
-      left: 32px;
-    }
-    .lds-roller div:nth-child(6) {
-      animation-delay: -0.216s;
-    }
-    .lds-roller div:nth-child(6):after {
-      top: 68px;
-      left: 24px;
-    }
-    .lds-roller div:nth-child(7) {
-      animation-delay: -0.252s;
-    }
-    .lds-roller div:nth-child(7):after {
-      top: 63px;
-      left: 17px;
-    }
-    .lds-roller div:nth-child(8) {
-      animation-delay: -0.288s;
-    }
-    .lds-roller div:nth-child(8):after {
-      top: 56px;
-      left: 12px;
-    }
-    @keyframes lds-roller {
-      0% {
-        transform: rotate(0deg);
-      }
-      100% {
-        transform: rotate(360deg);
-      }
-    }
   }
 
   header {
